using module (instead of global) for helper functions
also moved jshint globals out of comments into .jshintrc
