+++ test/e2e/.jshintrc
@@ -1,0 +1,30 @@
+{
+  "node": true,
+  "bitwise": true,
+  "camelcase": true,
+  "curly": true,
+  "eqeqeq": true,
+  "immed": true,
+  "indent": 2,
+  "latedef": true,
+  "newcap": true,
+  "noarg": true,
+  "quotmark": "single",
+  "regexp": true,
+  "undef": true,
+  "unused": true,
+  "strict": true,
+  "trailing": true,
+  "smarttabs": true,
+  "globals": {
+    "angular": false,
+    "browser": false,
+    "protractor": false,
+    "beforeAll": false,
+    "describe": false,
+    "it": false,
+    "element": false,
+    "expect": false,
+    "by": false
+  }
+}
