+++ test/e2e/specs/set-basemap.js
@@ -1,5 +1,7 @@
-/* globals beforeAll, describe, it, browser, element, expect, by, waitUntilElementIsReady, getAsyncAttributeValue*/
 'use strict';
+
+var helper = require('../helper');
+
 describe('Set Basemap', function() {
     // shared element locators
     var map = element(by.id('map'));
@@ -12,14 +14,14 @@ describe('Set Basemap', function() {
     it('should click on the "zoom in" and change the map "data-zoom" value from "3" to "4"', function() {
         // element locator(s) specific to this test
         var zoomIn = element(by.css('.esriSimpleSliderIncrementButton'));
-        waitUntilElementIsReady(zoomIn);
-        waitUntilElementIsReady(map);
+        helper.waitUntilElementIsReady(zoomIn);
+        helper.waitUntilElementIsReady(map);
 
         expect(map.getAttribute('data-zoom')).toEqual('3');
 
         zoomIn.click();
 
-        getAsyncAttributeValue(map, 'data-zoom').then(function(newValue) {
+        helper.getAsyncAttributeValue(map, 'data-zoom').then(function(newValue) {
             expect(newValue).toEqual('4');
         });
     });
@@ -27,13 +29,13 @@ describe('Set Basemap', function() {
     it('should choose a different basemap select and change the map "data-basemap" value from "satellite" to "oceans"', function() {
         // element locator(s) specific to this test
         var basemapSelect = element(by.model('map.basemap'));
-        waitUntilElementIsReady(map);
+        helper.waitUntilElementIsReady(map);
 
         expect(map.getAttribute('data-basemap')).toEqual('satellite');
 
         basemapSelect.sendKeys('oceans');
 
-        getAsyncAttributeValue(map, 'data-basemap').then(function(newValue) {
+        helper.getAsyncAttributeValue(map, 'data-basemap').then(function(newValue) {
             expect(newValue).toEqual('oceans');
         });
     });
